requirements.txt
apache-airflow
pandas
scrapy
itemadapter
requests
openpyxl
sqlalchemy
tqdm
selenium
2captcha-python
python-dotenv
pdfplumber
PyPDF2
psycopg2-binary
psycopg2
dependency-injector
unicode
xlrd